requirements.txt
appdirs==1.4.4
appnope==0.1.3
arrow==1.2.2
astor==0.8.1
asttokens==2.0.5
attrs==21.2.0
backcall==0.2.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.9
cheroot==8.5.2
CherryPy==18.6.1
click==8.0.3
cycler==0.11.0
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.0
dnspython==2.1.0
et-xmlfile==1.1.0
executing==0.8.3
firebase==3.0.1
firebase-token-generator==2.0.1
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.3
gcloud==0.17.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-resources==5.4.0
ipython==8.2.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.2
jupyter-core==4.9.1
jws==0.1.3
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.12.0
multitasking==0.0.10
nbformat==5.1.3
numpy==1.21.4
oauth2client==3.0.0
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandasgui==0.2.13
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
portend==3.1.0
prompt-toolkit==3.0.29
protobuf==3.19.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pygments==2.11.2
pymongo==4.0.1
pynput==1.7.6
pyobjc-core==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Quartz==8.5
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
Pyrebase==3.0.27
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jwt==2.0.1
pytz==2021.3
qtstylish==0.1.5
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.29
sseclient==0.0.27
stack-data==0.2.0
tempora==4.1.2
tenacity==8.0.1
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wordcloud==1.8.1
zc.lockfile==2.0
zipp==3.6.0
